+++ CHANGELOG.md
@@ -6,8 +6,19 @@ comprehensive report of every change we made in a release, it does provide detai
 might need to be aware of.
 
 - [Changes in Ice 3.9.0](#changes-in-ice-390)
+  - [C# Changes](#c-changes)
 
 ## Changes in Ice 3.9.0
+
+### C# Changes
+
+- Fixed a bug in `slice2cs --icerpc` where the generated request decoder read in-parameters in declaration order
+  instead of the marshal order. As a result, an operation with an optional parameter declared before a required
+  parameter (or optional parameters declared out of tag order) could fail to decode or decode incorrect values.
+
+- Fixed a bug in `slice2cs --icerpc` where the generated proxy built the response tuple in marshal order while
+  declaring it in declaration order. As a result, an operation whose out-parameters were declared in an order
+  different from their marshal order could return values in the wrong tuple slots or fail to compile.
 
 These are the changes since the [Ice 3.8.2] release.